Backend/requirements.txt
Flask>=2.2
redis>=5.0
pymongo>=4.6
openai>=1.0
python-dotenv>=1.0
sentence-transformers>=2.2.2
numpy>=1.24
bcrypt>=4.1
tiktoken>=0.6
prompt_toolkit>=3.0